Modify CCFontDefinition.FontSize to be float instead of int.  Remove casting to int in all cases and let the native interface handle the differences with the float size.
See issue https://github.com/mono/CocosSharp/issues/270

+++ src/label_nodes/CCLabel-CoreGraphics.cs
@@ -182,7 +182,7 @@ namespace CocosSharp
             var textDef = textDefinition;
             var contentScaleFactorWidth = CCLabel.DefaultTexelToContentSizeRatios.Width;
             var contentScaleFactorHeight = CCLabel.DefaultTexelToContentSizeRatios.Height;
-            textDef.FontSize *= (int)contentScaleFactorWidth;
+            textDef.FontSize *= contentScaleFactorWidth;
             textDef.Dimensions.Width *= contentScaleFactorWidth;
             textDef.Dimensions.Height *= contentScaleFactorHeight;
 
@@ -405,7 +405,7 @@ namespace CocosSharp
             var textDef = textDefinition;
             var contentScaleFactorWidth = CCLabel.DefaultTexelToContentSizeRatios.Width;
             var contentScaleFactorHeight = CCLabel.DefaultTexelToContentSizeRatios.Height;
-            textDef.FontSize *= (int)contentScaleFactorWidth;
+            textDef.FontSize *= contentScaleFactorWidth;
             textDef.Dimensions.Width *= contentScaleFactorWidth;
             textDef.Dimensions.Height *= contentScaleFactorHeight;
