+++ src/label_nodes/CCLabel-Gdi.cs
@@ -303,7 +303,7 @@ namespace CocosSharp
             var textDef = textDefinition;
             var contentScaleFactorWidth = CCLabel.DefaultTexelToContentSizeRatios.Width;
             var contentScaleFactorHeight = CCLabel.DefaultTexelToContentSizeRatios.Height;
-            textDef.FontSize *= (int)contentScaleFactorWidth;
+            textDef.FontSize *= contentScaleFactorWidth;
             textDef.Dimensions.Width *= contentScaleFactorWidth;
             textDef.Dimensions.Height *= contentScaleFactorHeight;
